+++ lib/llm/providers/xai/images.rb
@@ -15,6 +15,11 @@ class LLM::XAI
   #   IO.copy_stream res.images[0], "rocket.png"
   class Images < LLM::OpenAI::Images
     ##
+    # @api private
+    PATTERN = %r{\A(?:https?://|data:)}
+    private_constant :PATTERN
+
+    ##
     # Create an image
     # @example
     #   llm = LLM.xai(key: ENV["KEY"])
@@ -36,9 +41,42 @@ class LLM::XAI
     end
 
     ##
-    # @raise [NotImplementedError]
-    def edit(model: "grok-imagine-image-quality", **)
-      raise NotImplementedError
+    # Edit an image
+    # @example
+    #   llm = LLM.xai(key: ENV["KEY"])
+    #   res = llm.images.edit(image: "/images/book.png", prompt: "The book is floating in the clouds")
+    #   IO.copy_stream res.images[0], "floating-book.png"
+    # @see https://docs.x.ai/docs/guides/image-generations xAI docs
+    # @param [String, LLM::File, File] image The image to edit
+    # @param [String] prompt The prompt
+    # @param [String] model The model to use
+    # @param [Hash] params Other parameters (see xAI docs)
+    # @raise (see LLM::Provider#request)
+    # @return [LLM::Response]
+    def edit(image:, prompt:, model: "grok-imagine-image-quality", **params)
+      req = LLM::Transport::Request.post(path("/images/edits"), headers)
+      req.body = LLM.json.dump({
+        prompt:,
+        model:,
+        image: image_url(image),
+        response_format: "b64_json"
+      }.merge!(params))
+      res, span, tracer = execute(request: req, operation: "request")
+      res = LLM::OpenAI::ResponseAdapter.adapt(res, type: :image)
+      tracer.on_request_finish(operation: "request", model:, res:, span:)
+      res
+    end
+
+    private
+
+    def image_url(image)
+      case image
+      when String
+        url = image.match?(PATTERN) ? image : LLM.File(image).to_data_uri
+      else
+        url = LLM.File(image).to_data_uri
+      end
+      {url:, type: "image_url"}
     end
   end
 end